Add no-IPFS version of the network subgraph
