+++ config/ipfs.json
@@ -1,0 +1,3 @@
+{
+  "ipfs": true
+}
